Fix: Correct MVVM directory structure
This commit addresses your feedback on the MVVM refactoring.
The `MyHomePage` widget (View component) has been moved from `lib/main.dart` into its own file `lib/view/home_page.dart` as per standard MVVM practice.

Changes include:
- I created `lib/view/home_page.dart` and moved `MyHomePage` into it.
- I updated `lib/main.dart` to remove `MyHomePage` and import it from the new location.
- I ensured all relevant files (`main.dart`, `counter_model.dart`, `counter_viewmodel.dart`, `home_page.dart`) use consistent package imports.
- I verified widget tests in `test/widget_test.dart` remain functional with the new structure.

The `lib` directory now more accurately reflects the MVVM pattern with distinct `model`, `view`, and `viewmodel` directories containing their respective components.
