+++ lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart'; // Import provider
-import './viewmodel/counter_viewmodel.dart'; // Import CounterViewModel
+import 'package:provider/provider.dart';
+import 'package:flutter_app/viewmodel/counter_viewmodel.dart'; // Ensure this is a package import
+import 'package:flutter_app/view/home_page.dart'; // Import the new home_page.dart
 
 void main() {
   runApp(const MyApp());
@@ -11,52 +12,14 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return ChangeNotifierProvider( // Wrap MaterialApp with ChangeNotifierProvider
+    return ChangeNotifierProvider(
       create: (context) => CounterViewModel(),
       child: MaterialApp(
         title: 'Flutter Demo',
         theme: ThemeData(
           colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
         ),
-        home: const MyHomePage(title: 'Flutter Demo Home Page'),
-      ),
-    );
-  }
-}
-
-class MyHomePage extends StatelessWidget { // Change to StatelessWidget
-  const MyHomePage({super.key, required this.title});
-
-  final String title;
-
-  @override
-  Widget build(BuildContext context) {
-    // Get the CounterViewModel from the Provider
-    final counterViewModel = Provider.of<CounterViewModel>(context);
-
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-        title: Text(widget.title),
-      ),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            const Text('You have pushed the button this many times:'),
-            Text(
-              // Display the counter value from the CounterViewModel
-              '${counterViewModel.count}',
-              style: Theme.of(context).textTheme.headlineMedium,
-            ),
-          ],
-        ),
-      ),
-      floatingActionButton: FloatingActionButton(
-        // Call the incrementCounter method from the CounterViewModel
-        onPressed: () => counterViewModel.incrementCounter(),
-        tooltip: 'Increment',
-        child: const Icon(Icons.add),
+        home: const MyHomePage(title: 'Flutter Demo Home Page'), // This remains the same
       ),
     );
   }
