+++ lib/viewmodel/counter_viewmodel.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/foundation.dart';
-import '../model/counter_model.dart';
+import 'package:flutter_app/model/counter_model.dart'; // Changed to package import
 
 class CounterViewModel extends ChangeNotifier {
   CounterModel _counter = CounterModel();
